Desktop/uofthacks/apify-twitter-python/requirements.txt
apify-client
google-genai
python-dotenv
